requirements.txt
requests
psycopg2-binary
pandas
sqlalchemy
streamlit